+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added Mastodon API endpoint for updating media descriptions.
+- Add `name` attribute to media attachment if it has description.
 
 ### Changed